refactor: rename collectionStore to appStore
The file contained the full unified app state (collections, variables,
environments, responses, view state, app settings) but was still named
after the original collection-only store. Rename to appStore to match
the internal AppStoreState type, AppStore alias, and useAppStore hook.
