+++ src/renderer/state/CollectionStoreProvider.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, act, waitFor } from '@testing-library/react';
 import { CollectionStoreProvider } from './CollectionStoreProvider';
-import { AppStoreState, useAppStore } from './collectionStore';
+import { AppStoreState, useAppStore } from './appStore';
 
 const listeners: Record<string, (...args: unknown[]) => void> = {};
