+++ src/renderer/view/RequestWindow.tsx
@@ -1,6 +1,6 @@
 import { MainTopBar } from '@/components/mainWindow/MainTopBar';
 import { MainBody } from '@/components/mainWindow/MainBody';
-import { useCollectionActions, useCollectionStore } from '@/state/collectionStore';
+import { useCollectionActions, useCollectionStore } from '@/state/appStore';
 import { EmptyWildWest } from '@/assets/EmptyWildWest';
 import { MouseEvent, useCallback, useEffect, useRef } from 'react';
 import { registerGetRequest } from '@/lib/monaco/models';
